requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-ckeditor==5.7.1
django-extensions==2.2.1
django-js-asset==1.2.2
django-nested-inline==0.3.7
django-oauth-toolkit==1.0.0
djangorestframework==3.10.2
idna==2.8
oauthlib==2.0.3
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
